requirements.txt
pypdf2==3.0.1
python-dotenv==1.0.0
streamlit==1.33.0
google-cloud-aiplatform
google-generativeai